FlaskServer/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.18.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.1
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
billiard==4.2.4
blinker==1.9.0
cachetools==5.5.2
celery==5.6.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cyclopts==3.22.5
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
dotenv==0.9.9
ecdsa==0.19.1
edge-tts==7.2.3
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.124.0
fastmcp==2.11.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kombu==5.6.1
lazy-object-proxy==1.11.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.91.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
parse==1.20.2
passlib==1.7.4
pathable==0.4.4
pillow==11.3.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
postgrest==2.25.0
prisma==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyperclip==1.9.0
pytesseract==0.3.13
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python3-openid==3.2.0
PyYAML==6.0.2
qdrant-client==1.16.1
readabilipy==0.3.0
realtime==2.25.0
redis==7.1.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
soupsieve==2.7
SQLAlchemy==2.0.45
sqlparse==0.5.4
sse-starlette==3.0.2
starlette==0.47.2
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.235
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.12.8
